% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_setup.R
\name{add_units}
\alias{add_units}
\title{Add units to support unit conversion for calculations}
\usage{
add_units()
}
\value{
NULL, the function is used for its side-effects
}
\description{
Units are added to support molecular weight and mEq/L conversions.  Units are
named with the unit, an underscore, and the chemical species in lower case.
Examples are \code{"g_ammonia"}, \code{"mol_ammonia"}, and
\code{"Eq_ammonia"}.  Species with units are all species inputs for the
\code{equil2()} function.
}
\keyword{Internal}
